fix(ui): watch the heartbeat continuously instead of only on resume
The resume gate read the stream's last proof of life once, at the moment a
page came back, and then never again. A stream that died one second after
its last heartbeat still reported OPEN, so it kept its verdict until the
next reactivation -- which may never come while the tab stays open, leaving
a missed reply absent indefinitely (Codex P1 on #1597).

A heartbeat is a continuous clock, so give it a watchdog: arm it wherever a
stamp is taken, and on expiry either re-arm (a heartbeat landed) or declare
the stream dead. A hidden page still defers to the reactivation edge, since
its timers are throttled and it cannot hold a stream open anyway.

That makes onResumeGap redundant. wake() already recycles any stream that
cannot prove it survived, and every reconnect dispatches onConnected -- so
the gap signal was firing alongside it (a double catch-up on the zombie
path) and firing before the replacement subscription existed. Delete it:
onConnected is the one place every gap ends, dispatched once the new
subscription is live. WorkbenchInboxProvider gains it, which also closes the
window between its own first read and its subscription.

failWorkbenchEventStream() now owns the "this stream is over" transition, so
a silent death and a reported error leave exactly the same state behind.
